BASELINE: Update Chromium to 49.0.2623.23
Also adds missing printing sources.

+++ chromium/chrome_elf/blacklist/blacklist.cc
@@ -9,7 +9,6 @@
 
 #include <vector>
 
-#include "base/basictypes.h"
 #include "chrome_elf/blacklist/blacklist_interceptions.h"
 #include "chrome_elf/chrome_elf_constants.h"
 #include "chrome_elf/chrome_elf_util.h"
@@ -243,7 +242,9 @@ bool IsBlacklistInitialized() {
 }
 
 int GetBlacklistIndex(const wchar_t* dll_name) {
-  for (int i = 0; i < kTroublesomeDllsMaxCount && g_troublesome_dlls[i]; ++i) {
+  for (int i = 0;
+       i < static_cast<int>(kTroublesomeDllsMaxCount) && g_troublesome_dlls[i];
+       ++i) {
     if (_wcsicmp(dll_name, g_troublesome_dlls[i]) == 0)
       return i;
   }
@@ -253,7 +254,7 @@ int GetBlacklistIndex(const wchar_t* dll_name) {
 bool AddDllToBlacklist(const wchar_t* dll_name) {
   int blacklist_size = BlacklistSize();
   // We need to leave one space at the end for the null pointer.
-  if (blacklist_size + 1 >= kTroublesomeDllsMaxCount)
+  if (blacklist_size + 1 >= static_cast<int>(kTroublesomeDllsMaxCount))
     return false;
   for (int i = 0; i < blacklist_size; ++i) {
     if (!_wcsicmp(g_troublesome_dlls[i], dll_name))
