+++ chromium/sandbox/win/src/handle_closer_agent.cc
@@ -4,6 +4,9 @@
 
 #include "sandbox/win/src/handle_closer_agent.h"
 
+#include <limits.h>
+#include <stddef.h>
+
 #include "base/logging.h"
 #include "sandbox/win/src/nt_internals.h"
 #include "sandbox/win/src/win_utils.h"
@@ -120,7 +123,7 @@ void HandleCloserAgent::InitializeHandlesToClose(bool* is_csrss_connected) {
 
     DCHECK(reinterpret_cast<base::char16*>(entry) >= input);
     DCHECK(reinterpret_cast<base::char16*>(entry) - input <
-           sizeof(size_t) / sizeof(base::char16));
+           static_cast<ptrdiff_t>(sizeof(size_t) / sizeof(base::char16)));
   }
 
   // Clean up the memory we copied over.
